Move constants.ts to src root. Improve featherWeight test.
